feat: pre-ship v0.8.27 — macOS CI, trust docs, sign_in_active, server_personal
Ship-ready hardening before tag: macOS build_macos.sh + release.yml build-macos job, unsigned-beta trust notes, sign_in_active on update-check UI, personal doc extract to shared/server_personal.py, apply_update.sh without python3, runner module-size CI budget.
