+++ guide/getting-started.md
@@ -17,7 +17,7 @@ On Windows, use the project `.venv` Python - not the Microsoft Store `python.exe
 If you received a beta invite:
 
 1. Download and run **BAKLOG-Setup.exe** from your invite link.
-2. If SmartScreen warns about an unknown publisher, click **More info**, then **Run anyway**.
+2. If SmartScreen warns about an unknown publisher, click **More info**, then **Run anyway**. BAKLOG is an unsigned beta — see [Unsigned beta builds](troubleshooting.md#unsigned-beta-builds-no-code-signing).
 3. Launch **BAKLOG** from the Start Menu. A tray icon appears and your browser opens.
 4. Open the **Connections** tab and connect each store you use (Chrome or Edge required).
 
@@ -33,6 +33,13 @@ PORT=8766
 That keeps library files out of `BAKLOG-Data` and gives dev its own browser origin so localStorage is not shared with the installed app on port 8765. The dashboard header shows a **Dev server** chip when the dev server is active. See [troubleshooting](troubleshooting.md#dev-server-vs-frozen-exe-same-browser-origin).
 
 Portable zip builds work too: unzip to a normal folder, then run `Start BAKLOG.bat` or `BAKLOG Tray.exe`. Add an empty `portable.txt` beside `BAKLOG.exe` only if you want all data in the unzip folder (thumb drives). See `BETA-README.txt` in the bundle.
+
+## Beta (macOS)
+
+1. Download **BAKLOG-macos.zip** from [GitHub Releases](https://github.com/Ogrods/BAKLOG/releases/latest) (or your invite link once tagged).
+2. Unzip to a normal folder (not Downloads directly — Gatekeeper is stricter there). Double-click **Start BAKLOG.command** or run **BAKLOG Tray** from the `BAKLOG` folder.
+3. If macOS blocks the app ("cannot be opened"), right-click **BAKLOG Tray** → **Open** once, or run `xattr -cr /path/to/BAKLOG` in Terminal. See [Unsigned beta builds](troubleshooting.md#unsigned-beta-builds-no-code-signing).
+4. Library data lives in `~/Library/Application Support/BAKLOG-Data` (or beside the app when `portable.txt` is present). In-app updates use `apply_update.sh` once a mac zip is on the release.
 
 ## Install (from source)
