+++ scripts/check-module-size.mjs
@@ -11,6 +11,10 @@ const jsDir = path.join(root, 'js');
 const MAX_LINES = 3800;
 const EXEMPT = new Set(['table-query.worker.js']);
 
+/** Monolith runner — split incrementally; budget ratchets down after extractions. */
+const RUNNER_MAX_LINES = 2200;
+const RUNNER_FILES = ['js/fetcher/runner/index.js'];
+
 function lineCount(filePath) {
   return fs.readFileSync(filePath, 'utf8').split(/\r?\n/).length;
 }
@@ -21,15 +25,24 @@ for (const ent of fs.readdirSync(jsDir, { withFileTypes: true })) {
   if (EXEMPT.has(ent.name)) continue;
   const p = path.join(jsDir, ent.name);
   const lines = lineCount(p);
-  if (lines > MAX_LINES) offenders.push({ file: `js/${ent.name}`, lines });
+  if (lines > MAX_LINES) offenders.push({ file: `js/${ent.name}`, lines, max: MAX_LINES });
+}
+
+for (const rel of RUNNER_FILES) {
+  const p = path.join(root, rel);
+  if (!fs.existsSync(p)) continue;
+  const lines = lineCount(p);
+  if (lines > RUNNER_MAX_LINES) {
+    offenders.push({ file: rel, lines, max: RUNNER_MAX_LINES });
+  }
 }
 
 if (offenders.length) {
-  console.error(`Module size budget exceeded (max ${MAX_LINES} lines):`);
+  console.error('Module size budget exceeded:');
   for (const o of offenders.sort((a, b) => b.lines - a.lines)) {
-    console.error(`  ${o.file}: ${o.lines} lines`);
+    console.error(`  ${o.file}: ${o.lines} lines (max ${o.max})`);
   }
   process.exit(1);
 }
 
-console.log(`OK — all js/*.js modules are within ${MAX_LINES} lines.`);
+console.log(`OK — js/*.js ≤ ${MAX_LINES} lines; runner ≤ ${RUNNER_MAX_LINES}.`);
