+++ shared/install_visibility.py
@@ -63,10 +63,31 @@ def arp_version_mismatch(current_version: str, arp_version: str | None) -> bool:
     return normalize_version_tag(arp_version) != normalize_version_tag(current_version)
 
 
+def install_trust_fields() -> dict[str, Any]:
+    """Unsigned-beta trust notes for frozen builds (no code signing yet)."""
+    if not is_frozen():
+        return {}
+    fields: dict[str, Any] = {"unsigned_beta": True}
+    if sys.platform == "win32":
+        fields["trust_note"] = (
+            "Unsigned beta: SmartScreen may warn on first launch — More info, then Run anyway. "
+            "In-app zip updates do not refresh Add/Remove Programs; re-run Setup when you want ARP to match."
+        )
+    elif sys.platform == "darwin":
+        fields["trust_note"] = (
+            "Unsigned beta: Gatekeeper may block after in-app update — right-click Open once, "
+            "or run: xattr -cr /path/to/BAKLOG.app (or the BAKLOG folder) in Terminal."
+        )
+    else:
+        fields["trust_note"] = "Unsigned beta build — verify downloads from GitHub Releases only."
+    return fields
+
+
 def install_visibility_fields(current_version: str) -> dict[str, Any]:
     arp = read_arp_display_version()
     return {
         "install_source": detect_install_source(),
         "arp_version": arp,
         "arp_version_mismatch": arp_version_mismatch(current_version, arp),
+        **install_trust_fields(),
     }
